requirements.txt
pyqrcode
pillow
requests
